+++ RvR/math/RvR_rand_xor.c
@@ -115,9 +115,9 @@ uint64_t RvR_rand_xor_next(RvR_rand_xor *xor)
    return x + y;
 }
 
-int32_t RvR_rand_xor_next_range(RvR_rand_xor *xor, int32_t min, int32_t max)
+int64_t RvR_rand_xor_next_range(RvR_rand_xor * xor, int64_t min, int64_t max)
 {
-   uint32_t range = RvR_abs((max - min)) + 1;
+   uint64_t range = RvR_abs((max - min)) + 1;
    return min + (RvR_rand_xor_next(xor) % range);
 }
